move the check*Request functions to restutils (as Form may be ditched soon); change all calls - check for id in POST there accordingly
