+++ src/Graviton/RestBundle/Service/RestUtils.php
@@ -5,11 +5,17 @@
 
 namespace Graviton\RestBundle\Service;
 
+use Graviton\ExceptionBundle\Exception\InvalidJsonPatchException;
+use Graviton\ExceptionBundle\Exception\MalformedInputException;
+use Graviton\ExceptionBundle\Exception\NoInputException;
 use Graviton\JsonSchemaBundle\Validator\Validator;
 use Graviton\RestBundle\Model\DocumentModel;
 use Graviton\SchemaBundle\SchemaUtils;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\Routing\Route;
 use Symfony\Component\Routing\Router;
 use JMS\Serializer\Serializer;
@@ -186,6 +192,103 @@ final class RestUtils implements RestUtilsInterface
         );
 
         return $this->schemaValidator->validate(json_decode($content), json_decode($schema));
+    }
+
+    /**
+     * validate raw json input
+     *
+     * @param Request       $request  request
+     * @param Response      $response response
+     * @param DocumentModel $model    model
+     * @param string        $content  Alternative request content.
+     *
+     * @return void
+     */
+    public function checkJsonRequest(Request $request, Response $response, DocumentModel $model, $content = '')
+    {
+        if (empty($content)) {
+            $content = $request->getContent();
+        }
+
+        if (is_resource($content)) {
+            throw new BadRequestHttpException('unexpected resource in validation');
+        }
+
+        // is request body empty
+        if ($content === '') {
+            $e = new NoInputException();
+            $e->setResponse($response);
+            throw $e;
+        }
+
+        $input = json_decode($content, true);
+        if (JSON_ERROR_NONE !== json_last_error()) {
+            $e = new MalformedInputException($this->getLastJsonErrorMessage());
+            $e->setErrorType(json_last_error());
+            $e->setResponse($response);
+            throw $e;
+        }
+        if (!is_array($input)) {
+            $e = new MalformedInputException('JSON request body must be an object');
+            $e->setResponse($response);
+            throw $e;
+        }
+
+        if ($request->getMethod() == 'PUT' && array_key_exists('id', $input)) {
+            // we need to check for id mismatches....
+            if ($request->attributes->get('id') != $input['id']) {
+                $e = new MalformedInputException('Record ID in your payload must be the same');
+                $e->setResponse($response);
+                throw $e;
+            }
+        }
+
+        if ($request->getMethod() == 'POST' &&
+            array_key_exists('id', $input) &&
+            !$model->isIdInPostAllowed()
+        ) {
+            $e = new MalformedInputException(
+                '"id" can not be given on a POST request. Do a PUT request instead to update an existing record.'
+            );
+            $e->setResponse($response);
+            throw $e;
+        }
+    }
+
+    /**
+     * Validate JSON patch for any object
+     *
+     * @param array $jsonPatch json patch as array
+     *
+     * @throws InvalidJsonPatchException
+     * @return void
+     */
+    public function checkJsonPatchRequest(array $jsonPatch)
+    {
+        foreach ($jsonPatch as $operation) {
+            if (!is_array($operation)) {
+                throw new InvalidJsonPatchException('Patch request should be an array of operations.');
+            }
+            if (array_key_exists('path', $operation) && trim($operation['path']) == '/id') {
+                throw new InvalidJsonPatchException('Change/remove of ID not allowed');
+            }
+        }
+    }
+
+    /**
+     * Used for backwards compatibility to PHP 5.4
+     *
+     * @return string
+     */
+    private function getLastJsonErrorMessage()
+    {
+        $message = 'Unable to decode JSON string';
+
+        if (function_exists('json_last_error_msg')) {
+            $message = json_last_error_msg();
+        }
+
+        return $message;
     }
 
     /**
